Move utils/* into bundle_tools/
as bundle_tools is it's only user.

Delete obsolete s3_examples/
