+++ bin/bundle_tools/__init__.py
@@ -2,3 +2,4 @@ from .bundle import Bundle, LocalBundle, StagedBundle, File, MetadataFile, DataF
 from .submission import SubmissionInfo
 from .staging import BundleStager
 from .storage import StagedBundleFinder, BundleStorer, DataStoreAPI
+from .parallel_logger import logger
