--- bin/s3_examples/__init__.py
@@ -1,6 +1,0 @@
-"""
-Classes that interact with the example data bundles in S3
-"""
-
-from .s3_example_bundle import S3ExampleBundle
-from .s3_example_file import S3ExampleFile
